不要なモデルファイルを削除し、プロフィールコンポーネントを追加して「About」ページを実装
