--- model/jsonApi.model.ts
@@ -1,7 +1,0 @@
-export interface JsonApi<T> {
-  data: T;
-}
-
-export interface JsonApiRef<T, U> extends JsonApi<T> {
-  included: U[];
-}
